+++ core/model/src/main/java/cn/qihangerp/model/entity/OOrderItem.java
@@ -132,9 +132,9 @@ public class OOrderItem implements Serializable {
     private Integer orderStatus;
 
     /**
-     * 发货方式 0 自己发货1联合发货2供应商发货
+     * 发货人
      */
-    private Integer shipper;
+    private Long shipper;
 
     /**
      * 发货方式 0 自己发货或待处理2供应商发货
